requirements-monitor.txt
annotated-types==0.7.0
anyio==4.9.0
apsw==3.49.1.0
apswutils==0.0.2
atproto==0.0.61
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
fastcore==1.8.1
fastlite==0.1.3
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
libipld==3.0.1
oauthlib==3.2.2
packaging==25.0
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-fasthtml==0.12.14
python-multipart==0.0.20
pyyaml==6.0.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==13.1
